Move classes to proper location; Delete unused imports
