+++ core/src/com/github/lhrb/nemo/actors/EnemyActor.java
@@ -8,10 +8,10 @@ import java.util.Random;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.actions.Actions;
 import com.github.lhrb.nemo.GameManager;
-import com.github.lhrb.nemo.SpawnFactory.CollectibleFactory;
-import com.github.lhrb.nemo.actors.enemies.Uboot;
 import com.github.lhrb.nemo.actors.shots.Shots;
 import com.github.lhrb.nemo.util.AnimationLoader;
+import com.github.lhrb.nemo.util.CollectibleFactory;
+import com.github.lhrb.nemo.util.CollisionEvent;
 import com.github.lhrb.nemo.util.SoundManager;
