+++ core/src/com/github/lhrb/nemo/actors/Existence.java
@@ -3,6 +3,8 @@
  */
 package com.github.lhrb.nemo.actors;
 
+import com.github.lhrb.nemo.util.CollisionEvent;
+
 /**
  * @author exa
  *
