+++ core/src/com/github/lhrb/nemo/screen/FirstLevelScreen.java
@@ -4,7 +4,7 @@ package com.github.lhrb.nemo.screen;
 import com.github.lhrb.nemo.actors.Player;
 import com.github.lhrb.nemo.ui.HUD;
 import com.github.lhrb.nemo.actors.Background;
-import com.github.lhrb.nemo.SpawnFactory.EnemyFactory;
+import com.github.lhrb.nemo.util.EnemyFactory;
 import com.github.lhrb.nemo.util.SoundManager;
 
 import java.beans.PropertyChangeSupport;
