+++ core/src/com/github/lhrb/nemo/util/AnimationLoader.java
@@ -13,7 +13,6 @@ import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.Array;
 
 import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.util.HashMap;
 
 /**
