+++ core/src/com/github/lhrb/nemo/util/CollectibleFactory.java
@@ -1,4 +1,4 @@
-package com.github.lhrb.nemo.SpawnFactory;
+package com.github.lhrb.nemo.util;
 
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.github.lhrb.nemo.actors.powerups.CType;
@@ -8,7 +8,7 @@ import java.util.Random;
 
 public class CollectibleFactory {
     //SEED 111 FOR TESTING
- private static Random rand = new Random();
+    private static Random rand = new Random();
 
     public CollectibleFactory() {
     }
@@ -16,8 +16,6 @@ public class CollectibleFactory {
 
     public static void spawnC(float x, float y, Stage stage) {
         
-        Random rand = new Random();
-        //if ((rand.nextInt(10)+1) < 9) {
         CType type;
 
             switch (rand.nextInt(8) + 1) {
